Refactor : popup 컴포넌트 폴더로 이동
